refactor(ai): consolidate trace support files
